requirements.txt
streamlit
python-dotenv
PyPDF2
langchain
langchain_community
faiss-cpu
huggingface_hub
htmlTemplates
